+++ editor/model/actions/SetGraphTensionAction.cpp
@@ -6,13 +6,13 @@
 
 using namespace AxiomModel;
 
-SetGraphTensionAction::SetGraphTensionAction(const QUuid &controlUuid, uint8_t index, float oldTension,
-                                             float newTension, AxiomModel::ModelRoot *root)
+SetGraphTensionAction::SetGraphTensionAction(const QUuid &controlUuid, uint8_t index, double oldTension,
+                                             double newTension, AxiomModel::ModelRoot *root)
     : Action(ActionType::SET_GRAPH_TENSION, root), _controlUuid(controlUuid), _index(index), _oldTension(oldTension),
       _newTension(newTension) {}
 
 std::unique_ptr<SetGraphTensionAction> SetGraphTensionAction::create(const QUuid &controlUuid, uint8_t index,
-                                                                     float oldTension, float newTension,
+                                                                     double oldTension, double newTension,
                                                                      AxiomModel::ModelRoot *root) {
     return std::make_unique<SetGraphTensionAction>(controlUuid, index, oldTension, newTension, root);
 }
